step 3: Move filterby ts to its own file
